billing_sqlite_arango/requirements.txt

python-arango>=7.0
reportlab
